Remove shape checks for split - XNNpack can now handle this.
PiperOrigin-RevId: 612423885

+++ src/xnnpack/subgraph.h
@@ -249,7 +249,7 @@ struct xnn_node {
   /// Static parameters of the operator node.
   union {
     struct {
-      size_t axis;
+      int32_t axis;
     } concatenate;
     struct {
       uint32_t input_padding_top;
@@ -427,7 +427,7 @@ struct xnn_operator_data {
       size_t reshape_dims[XNN_MAX_TENSOR_DIMS];
     };
     // Used for concatenate.
-    size_t axis;
+    int32_t axis;
     // Used for static constant pad.
     struct {
       size_t post_paddings[XNN_MAX_TENSOR_DIMS];
